backend/requirements.txt
# Core Framework
django~=5.1
djangorestframework~=3.15
python-dotenv~=1.0

# Authentication
firebase-admin~=6.5

# Database
psycopg2-binary~=2.9
dj-database-url~=2.2

# API & Middleware
django-cors-headers~=4.5
gunicorn~=22.0
whitenoise~=6.7

# AI / ML
numpy~=1.26
pandas~=2.2
scikit-learn~=1.5
google-generativeai~=0.8
deep-translator~=1.11
groq~=0.11
torch~=2.4
joblib~=1.4
